gh.js GitHub requires NoReferer
I really do not understand what is going on there.

For me it seems to be some security provisioning going in the wrong direction.
AFAICS this is only blocking legitimate links, but does not block attacks.
If it is so easy to circumvent it with setting rel="noreferrer" on the link
then something definitively is wrong or even worse.